Feat: Reorganize cache
